refactor: 공통적으로 사용될 타입 분리
